Adding flags

+++ cli/run.sh
@@ -2,6 +2,10 @@
 shortfile="$(basename $1)"
 name="${shortfile%.*}"
 
-rm cli/tmp/*
-bash cli/build.sh $@
-"./cli/tmp/${name}"
+# Check if directory isn't empty
+if [ $(ls -A cli/tmp | wc -l ) -ne 0 ]; then 
+    rm cli/tmp/*
+fi
+
+bash cli/build.sh $@ || exit 1
+"./cli/tmp/${name}" || exit 1
